Test: pass tests for disable-http1 and disable-nghttp2

+++ test/test_restconf_yang_patch_json.sh
@@ -8,9 +8,11 @@
 s="$_" ; . ./lib.sh || if [ "$s" = $0 ]; then exit 0; else return 0; fi
 
 # Enable if YANG_PATCH
-echo "...skipped: YANG_PATCH JSON NYI"
-rm -rf $dir
-if [ -z "${CLIXON_YANG_PATCH}" -a "$s" = $0 ]; then exit 0; else return 0; fi
+if true; then
+    echo "...skipped: YANG_PATCH JSON NYI"
+    rm -rf $dir
+    if [ -z "${CLIXON_YANG_PATCH}" -a "$s" = $0 ]; then exit 0; else return 0; fi
+fi
 
 APPNAME=example
